düzen

+++ BankaManagementSystem/BankaManagementSystem/Hesap_MüşteriHesabı.cs
@@ -13,7 +13,7 @@ namespace BankaManagementSystem
 {
     public partial class Hesap_MüşteriHesabı : Form
     {
-        public long tc;//kullanıcı bilgilerini formlar arası aktarmak için
+        public int tc;//kullanıcı bilgilerini formlar arası aktarmak için
         private Button currentButton;//aktif butonu tutmak için
         private Form activeForm;//form içerisine çağırılan formu tutmak için
