+++ BankaManagementSystem/BankaManagementSystem/Menu_Müşteri/TransferVeÖdemeler.cs
@@ -12,7 +12,7 @@ namespace BankaManagementSystem.Menu_Müşteri
 {
     public partial class TransferVeÖdemeler : Form
     {
-        public long tc;
+        public int tc;
         public TransferVeÖdemeler()
         {
             InitializeComponent();
